upload video withou physical video

+++ apps/upload/migrations/0019_calculate_file_size.py
@@ -14,29 +14,39 @@ def calculate_size(apps, schema_editor):
     images = []
     for obj in UploadFile.objects.all():
         try:
-            obj.file_size = ceil(obj.file_path.size / 1024)
-            files.append(obj)
+            if obj.file_size:
+                obj.file_size = ceil(obj.file_path.size / 1024)
+                files.append(obj)
         except Exception:
             obj.delete()
 
     for obj in UploadVideo.objects.all():
         try:
-            obj.video_size = ceil(obj.video_path.size / 1024)
-            videos.append(obj)
+            if obj.video_size:
+                obj.video_size = ceil(obj.video_path.size / 1024)
+                videos.append(obj)
         except Exception:
-            obj.delete()
+            if obj.video_embedded_url:
+                obj.video_size = None
+                obj.video_path = None
+                obj.video_type = None
+                obj.duration = None
+                videos.append(obj)
+            else:
+                obj.delete()
 
     for obj in UploadImage.objects.all():
         try:
-            obj.image_size = ceil(obj.image_path.size / 1024)
-            images.append(obj)
+            if obj.image_size:
+                obj.image_size = ceil(obj.image_path.size / 1024)
+                images.append(obj)
         except Exception:
             obj.delete()
 
     if files:
         UploadFile.objects.bulk_update(files, fields=["file_size"])
     if videos:
-        UploadVideo.objects.bulk_update(videos, fields=["video_size"])
+        UploadVideo.objects.bulk_update(videos, fields=["video_size", "video_path", "video_type", "duration"])
     if images:
         UploadImage.objects.bulk_update(images, fields=["image_size"])
